fix: cast the action type to float

+++ model.py
@@ -82,6 +82,6 @@ class Critic(nn.Module):
         xs = F.relu(self.fcs1(state))
         # applying a batch Normalization on the first layer output
         xs = self.bn1(xs)
-        x = torch.cat((xs, action), dim=1)
+        x = torch.cat((xs, action.float()), dim=1)
         x = F.relu(self.fc2(x))
         return self.fc3(x)
